refactor: organizing files & folders
