+++ react-game/src/App.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from './components/App/App';
 
 test('renders learn react link', () => {
   render(<App />);
