+++ react-game/src/components/App/App.scss
