+++ react-game/src/components/App/App.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
-import Footer from './Footer';
+import Footer from '../Footer/Footer';
+import Game from '../Game/Game';
 import './App.scss';
 
 const App: React.FC = () => {
   return (
     <div className="App">
+      <Game />
       <Footer />
     </div>
   );
