+++ react-game/src/components/Board/Board.scss
@@ -1,0 +1,5 @@
+.Board {
+  width: 600px;
+  height: 600px;
+  background-color: #eee;
+}
