+++ react-game/src/components/Board/Board.tsx
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import './Board.scss';
+
+class Board extends Component {
+  render() {
+    return <div className="Board" />;
+  }
+}
+
+export default Board;
