+++ react-game/src/components/Footer/Footer.scss
