+++ react-game/src/components/Game/Game.tsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import Board from '../Board/Board';
+import './Game.scss';
+
+class Game extends Component {
+  render() {
+    return (
+      <div>
+        <Board />
+      </div>
+    );
+  }
+}
+
+export default Game;
