+++ react-game/src/index.scss
@@ -1,4 +1,19 @@
+* {
+	margin: 0;
+	padding: 0;
+	box-sizing:border-box;
+}
+
+html {
+	width: 100%;
+	height: 100%;
+}
+
 body {
-  margin: 0;
-  height: 100vh;
+  width: 100%;
+	min-height: 100%;
+	display: flex;
+  flex-direction: column;
+	justify-content: center;
+	align-items: center;
 }
